+++ src/MeasureMap/SessionHandlers/BasicSessionHandler.cs
@@ -21,8 +21,14 @@ namespace MeasureMap
             
             var thread = runnerThreads.StartNew(0, _ =>
             {
+                var ctx = settings.OnStartPipeline();
+
                 var worker = new Worker();
-                return worker.Run(task, settings.OnStart());
+                var result = worker.Run(task, ctx);
+
+                settings.OnEndPipeline(ctx);
+
+                return result;
             }, settings.GetThreadFactory());
             
             settings.Logger.Write($"Start thread {thread.Id}", LogLevel.Debug, nameof(BasicSessionHandler));