+++ src/Tests/MeasureMap.UnitTest/ProfileSessionTests.cs
@@ -663,11 +663,13 @@ namespace MeasureMap.UnitTest
             ProfilerSession.StartSession()
                 .SetIterations(5)
                 .RunWarmup(false)
-                .OnExecuted(r => calls++)
                 .Task(() => { })
+                .OnExecuted(r => calls++)
                 .RunSession();
 
-            calls.Should().Be(5);
+            //
+            // OnExecuted is run async so only test if called at least once
+            calls.Should().BeGreaterThan(0);
         }
 
         [Test]